memcg: increase MEMCG_CHARGE_BATCH to 128
MEMCG_CHARGE_BATCH was last changed to 64 back in 2022, systems have
grown in memory and speed and it's useful to increase this to 128.

Benchmarking the stress-ng mmap stressor shows a performance improvement
of ~7.4% and malloc stressor by 2.8%, tested on an Ultra 9 285K with
turbo disabled to avoid test result jitter.

Signed-off-by: Colin Ian King <colin.i.king@intel.com>