soc: qcom: smp2p: Kick after requesting interrupt
There is a small chance that an already running remoteproc does not take
kicking it lightly and kicks us back immediately after. If this happens
during probe() there is an even smaller chance that we might miss the
incoming interrupt, because it is not registered yet.

Avoid this race condition by kicking the outgoing edge only after we
started listening for incoming interrupts.

Fixes: 50e99641413e ("soc: qcom: smp2p: Qualcomm Shared Memory Point to Point")
Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>
Signed-off-by: Abel Vesa <abel.vesa@oss.qualcomm.com>