arm64: dts: qcom: x1-el2: Add qcom,broken-reset for remoteprocs
The remoteproc PAS firmware interface does not work properly when running
bare-metal without hypervisor in EL2. It's possible to authenticate and
start new firmware, but the remoteproc is never actually brought out of
reset. On X1E this affects all remoteprocs: ADSP and CDSP.

Mark both with "qcom,broken-reset" to make the driver skip trying to reset
the remoteproc. We can still attach to them if they are already running
during boot. On X1E, the "lite" ADSP firmware should be always running
already during boot, and it provides at least battery status/charging as
well as USB-C detection for external displays.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
Signed-off-by: Abel Vesa <abel.vesa@oss.qualcomm.com>