drm/amd/display: Fix high busy wait load in dmub_srv_wait_for_idle()
dmub_srv_wait_for_idle() is called with a timeout_us of 100000, which is
100 ms, and loops with udelay(1) on every iteration.

Since the udelay duration is very short, the CPU doesn't have enough time
to enter a C-state to save power during the delay.

Worse, since the timeout limit is 100 ms, this permits the CPU to burn a
lot of time on busy waiting. The load from busy waiting is high enough that
it occasionally prevents the CPU from entering deeper implicit idle states
temporarily (package C-states), sometimes for one minute or more.

Empirically on a Strix Halo system (AMD RYZEN AI MAX+ PRO 395 w/ Radeon
8060S, HP ZBook Ultra G1a) it is observed that the total delayed time
consistently oscillates between a few different durations: 0-3 us,
10-30 us, and 100-1000+ us. Knowing this, implement progressively longer
delays following this pattern instead of delaying for 1 us every iteration.

Furthermore, when the context permits it, use usleep_range() instead of
busy waiting in order to reclaim CPU time lost to waiting here. This lost
CPU time is significant and amounts to ~5% of CPU cycles spent in the
kernel when the system is quiet, according to perf top.

This resolves the temporary power consumption spikes from the loss of
package C-state residency, and dmub_srv_wait_for_idle() no longer appears
in perf top.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Eric Naim <dnaim@cachyos.org>