rpmsg: core: Make it easier to manually create endpoint during probe()
Currently, the rpmsg core automatically creates an endpoint for rpmsg
drivers that specify a receive callback. This works "somewhat" in most
simple cases, but it is often prone to race conditions: The receive
callback can be called as soon and as long as the endpoint is open, so
drivers must be prepared to handle calls to the receive callback:

 - Before their probe() function is called
   (after the endpoint was created)
 - In parallel to their probe() function
 - In parallel to their remove() function
 - After their remove() function is called
   (before the endpoint is destroyed)

It is difficult for drivers to handle this without being able to run code
before endpoint creation and after endpoint destruction. Also, they may
need to hold locks while creating/destroying the endpoint to handle edge
cases reliably.

Drivers can already create endpoints manually by omitting the receive
callback in rpmsg_driver, but for most simple cases where there is only
a single channel, the boilerplate required for that is a bit cumbersome.

Add a rpmsg_dev_open_ept() function that can be called by drivers during
the probe() function. It results in effectively the same that the rpmsg
core would normally do if the receive callback is specified.
announce_create() and announce_destroy() are still handled by the rpmsg
core. During remove(), the drivers can directly call rpmsg_destroy_ept().

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>
Signed-off-by: Abel Vesa <abel.vesa@oss.qualcomm.com>